pyproject.toml
[tool.poetry]
name = "besafe"
version = "0.1.0"
description = ""
authors = ["HansolLim <dev.hansollim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.3"
django-mysql = "^4.10.0"
python-dotenv = "^1.0.0"
pymysql = "^1.1.0"
whitenoise = "^6.5.0"
setuptools = "^68.0.0"
gunicorn = "^19.1.1"
dj-soft-delete = "^1.0.1"
django-admin-ordering = "^0.17.0"
jinja2-django-tags = "^0.5"
pillow = "^10.1.0"
django-tinymce = "^3.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
